requirements.txt
asgiref==3.7.2
black==24.8.0
click==8.1.7
colorama==0.4.6
Django==5.0.2
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.7
Faker==28.4.1
flake8==7.1.1
inflection==0.5.1
Markdown==3.7
mccabe==0.7.0
mypy-extensions==1.0.0
mysqlclient==2.2.4
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pycodestyle==2.12.1
pyflakes==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requirements.py==1.0.1
setuptools==74.1.2
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
validate-docbr==1.10.0
